64bit compatibility fixes from Mike Gist.  Should fix crashing in certain animations on 64 bit platforms.

+++ cal3d/src/cal3d/coreanimation.cpp
@@ -215,7 +215,7 @@ unsigned int CalCoreAnimation::getTrackCount() const
   *
   *****************************************************************************/
 
-void CalCoreAnimation::setTrackAssignments(const std::vector<int>& trackAssignments)
+void CalCoreAnimation::setTrackAssignments(const std::vector<intptr_t>& trackAssignments)
 {
   m_skeletonMapping.resize(trackAssignments.size());
   for (unsigned index = 0; index < trackAssignments.size(); ++index)
